requirements.txt
absl-py==0.12.0
aiohttp==3.7.4.post0
argon2-cffi==20.1.0
async-generator==1.10
async-timeout==3.0.1
attrs==20.3.0
axial-positional-embedding==0.2.1
backcall==0.2.0
bleach==3.3.0
blobfile==1.1.2
brotlipy==0.7.0
cachetools==4.2.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
clip==1.0
cloudpickle==1.6.0
configparser==5.0.2
cryptography==3.4.7
cycler==0.10.0
cytoolz==0.11.0
DALL-E==0.1
dalle-pytorch==0.10.2
dask==2021.3.1
dataclasses==0.6
decorator==4.4.2
deepspeed==0.3.13+e721cb6
defusedxml==0.7.1
docker-pycreds==0.4.0
easydict==1.9
einops==0.3.0
entrypoints==0.3
filelock==3.0.12
fsspec==0.8.7
ftfy==5.9
future==0.18.2
gdown==3.12.2
gitdb==4.0.7
GitPython==3.1.14
google-auth==1.28.0
google-auth-oauthlib==0.4.4
grpcio==1.36.1
idna==2.10
imagecodecs==2021.1.11
imageio==2.9.0
iniconfig==1.1.1
ipykernel==5.3.4
ipython==7.18.1
ipython-genutils==0.2.0
ipywidgets==7.6.3
jedi==0.18.0
Jinja2==2.11.3
joblib==1.0.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.7
jupyter-console==6.4.0
jupyter-core==4.6.3
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
kiwisolver==1.3.1
locket==0.2.1
Markdown==3.3.4
MarkupSafe==1.1.1
matplotlib==3.3.4
mistune==0.8.4
mkl-fft==1.3.0
mkl-random==1.1.1
mkl-service==2.3.0
mpi4py==3.0.3
mpmath==1.2.1
multidict==5.1.0
mypy==0.812
mypy-extensions==0.4.3
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.1
networkx==2.5
ninja==1.10.0.post2
nltk==3.5
notebook==6.3.0
numpy==1.19.2
oauthlib==3.1.0
olefile==0.46
omegaconf==2.0.6
packaging==20.9
pandas==1.2.4
pandocfilters==1.4.3
parso==0.8.0
partd==1.2.0
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.1.2
pip==21.0.1
pluggy==0.13.1
prometheus-client==0.10.0
promise==2.3
prompt-toolkit==3.0.8
protobuf==3.15.6
psutil==5.8.0
ptyprocess==0.6.0
py==1.10.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pycryptodomex==3.10.1
Pygments==2.7.1
pyOpenSSL==20.0.1
pyparsing==2.4.7
pyrsistent==0.17.3
PySocks==1.7.1
pytest==6.2.2
python-dateutil==2.8.1
pytorch-lightning==1.2.6
pytz==2021.1
PyWavelets==1.1.1
PyYAML==5.3.1
pyzmq==19.0.2
qtconsole==5.0.3
QtPy==1.9.0
regex==2021.3.17
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
sacremoses==0.0.43
scikit-image==0.17.2
scipy==1.6.2
Send2Trash==1.5.0
sentry-sdk==1.0.0
setuptools==52.0.0.post20210125
shortuuid==1.0.1
six==1.15.0
smmap==4.0.0
subprocess32==3.5.4
sympy==1.7.1
taming-transformers==0.0.1
tensorboard==2.4.1
tensorboard-plugin-wit==1.8.0
tensorboardX==1.8
terminado==0.9.4
testpath==0.4.4
tifffile==2021.3.17
tokenizers==0.10.2
toml==0.10.2
toolz==0.11.1
torch==1.7.1
torchmetrics==0.2.0
torchvision==0.8.2
tornado==6.1
tqdm==4.59.0
traitlets==5.0.5
transformers==4.4.2
triton==0.2.3
typed-ast==1.4.2
typing-extensions==3.7.4.3
urllib3==1.26.4
wandb==0.10.26
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
wheel==0.36.2
widgetsnbextension==3.5.1
xmltodict==0.12.0
yarl==1.6.3
